Move readFileAtURL func to util
Moving readFileAtURL from dataset.go to util.go
since it is used in multiple files.
